refactor: add CardCalculator class
- separate the calculation logic of the sum of cards into the Card Calculator class.
