+++ src/main/kotlin/blackjack/Round.kt
@@ -5,31 +5,5 @@ class Round(private val trump: Trump = Trump()) {
 
     companion object {
         const val TWENTY_ONE = 21
-
-        fun calculateSum(cards: List<Card>): Int {
-            var sum = 0
-            cards.filterNot { it.value == CardValue.ACE }.forEach {
-                sum += it.value.toInt(sum)
-            }
-            cards.filter { it.value == CardValue.ACE }.forEach {
-                sum += it.value.toInt(sum)
-            }
-            return sum
-        }
-
-        private fun CardValue.toInt(sum: Int): Int {
-            return when (this) {
-                CardValue.TWO -> 2
-                CardValue.THREE -> 3
-                CardValue.FOUR -> 4
-                CardValue.FIVE -> 5
-                CardValue.SIX -> 6
-                CardValue.SEVEN -> 7
-                CardValue.EIGHT -> 8
-                CardValue.NINE -> 9
-                CardValue.ACE -> if (sum + 11 > TWENTY_ONE) 1 else 11
-                else -> 10
-            }
-        }
     }
 }
